core/requirements.txt
amqp==5.1.1
asgiref==3.2.10
astroid==2.11.7
async-timeout==4.0.2
billiard==3.6.4.0
black==22.6.0
celery==5.2.3
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dill==0.3.5.1
Django==3.1.5
django-celery-beat==2.3.0
django-cors-headers==3.11.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-timezone-field==5.0
djangorestframework==3.12.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
PyJWT==2.4.0
pylint==2.14.5
pylint-plugin-utils==0.7
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
